refactor(ui): improve entityinfoshort component
