+++ ui/components/compliance/compliance-header/compliance-scan-info.tsx
@@ -1,7 +1,7 @@
 import { Divider } from "@heroui/divider";
 import { Tooltip } from "@heroui/tooltip";
 
-import { DateWithTime, EntityInfoShort } from "@/components/ui/entities";
+import { DateWithTime, EntityInfo } from "@/components/ui/entities";
 import { ProviderType } from "@/types";
 
 interface ComplianceScanInfoProps {
@@ -20,15 +20,15 @@ interface ComplianceScanInfoProps {
 
 export const ComplianceScanInfo = ({ scan }: ComplianceScanInfoProps) => {
   return (
-    <div className="flex items-center gap-2">
-      <EntityInfoShort
-        cloudProvider={scan.providerInfo.provider}
-        entityAlias={scan.providerInfo.alias}
-        entityId={scan.providerInfo.uid}
-        hideCopyButton
-        snippetWidth="max-w-[100px]"
-      />
-      <Divider orientation="vertical" className="h-6" />
+    <div className="flex items-center gap-4">
+      <div className="flex shrink-0 items-center">
+        <EntityInfo
+          cloudProvider={scan.providerInfo.provider}
+          entityAlias={scan.providerInfo.alias}
+          entityId={scan.providerInfo.uid}
+        />
+      </div>
+      <Divider orientation="vertical" className="h-8" />
       <div className="flex flex-col items-start whitespace-nowrap">
         <Tooltip
           content={scan.attributes.name || "- -"}
