+++ ui/components/findings/table/finding-detail.tsx
@@ -14,7 +14,7 @@ import {
 } from "@/components/shadcn";
 import { CodeSnippet } from "@/components/ui/code-snippet/code-snippet";
 import { CustomLink } from "@/components/ui/custom/custom-link";
-import { EntityInfoShort, InfoField } from "@/components/ui/entities";
+import { EntityInfo, InfoField } from "@/components/ui/entities";
 import { DateWithTime } from "@/components/ui/entities/date-with-time";
 import { SeverityBadge } from "@/components/ui/table/severity-badge";
 import { buildGitFileUrl, extractLineRangeFromUid } from "@/lib/iac-utils";
@@ -119,7 +119,7 @@ export const FindingDetail = ({
         </CardHeader>
         <CardContent className="flex flex-col gap-4">
           <div className="flex flex-wrap gap-4">
-            <EntityInfoShort
+            <EntityInfo
               cloudProvider={providerDetails.provider as ProviderType}
               entityAlias={providerDetails.alias}
               entityId={providerDetails.uid}
