+++ ui/components/ui/entities/index.ts
@@ -1,5 +1,5 @@
 export * from "./date-with-time";
-export * from "./entity-info-short";
+export * from "./entity-info";
 export * from "./get-provider-logo";
 export * from "./info-field";
 export * from "./scan-status";
